Update: Scrpits CrossHairController and Weapon undergone changes to handling damage output
Replaced the use of instantiating projectile prefabs in world space with the use of the crosshair to send damage data to the current collider that the crosshair's collider overlaps with.
This may prevent extra work to perfect projectile movement and save space for unnecessary creation of more prefabs (projectiles) to cuter the world space.